Hospital_Management/hospitalproject/requirements.txt
django>=4.2
django-autoslug
razorpay
Pillow
mysqlclient
requests